(HDS-1854) Add timestamp update functionality support for window API call
